refactor(inputfield): add proper input molecule
